refactoring handlers
