+++ server/src/main/java/org/mskcc/cmo/metadb/MetadbApp.java
@@ -4,7 +4,9 @@ import java.util.concurrent.CountDownLatch;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.mskcc.cmo.messaging.Gateway;
-import org.mskcc.cmo.metadb.service.MessageHandlingService;
+import org.mskcc.cmo.metadb.service.ResearchMessageHandlingService;
+import org.mskcc.cmo.metadb.service.ClinicalMessageHandlingService;
+import org.mskcc.cmo.metadb.service.PatientCorrectionHandlingService;
 import org.mskcc.cmo.metadb.service.RequestReplyHandlingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
@@ -39,7 +41,13 @@ public class MetadbApp implements CommandLineRunner {
     private Gateway messagingGateway;
 
     @Autowired
-    private MessageHandlingService messageHandlingService;
+    private ResearchMessageHandlingService researchMessageHandlingService;
+    
+    @Autowired
+    private ClinicalMessageHandlingService clinicalMessageHandlingService;
+    
+    @Autowired
+    private PatientCorrectionHandlingService patientCorrectionHandlingService;
 
     @Autowired
     private RequestReplyHandlingService requestReplyHandlingService;
@@ -75,7 +83,9 @@ public class MetadbApp implements CommandLineRunner {
             installShutdownHook();
             messagingGateway.connect();
             requestReplyHandlingService.initialize(messagingGateway);
-            messageHandlingService.initialize(messagingGateway);
+            researchMessageHandlingService.initialize(messagingGateway);
+            clinicalMessageHandlingService.initialize(messagingGateway);
+            patientCorrectionHandlingService.initialize(messagingGateway);
             metadbAppClose.await();
         } catch (Exception e) {
             LOG.error("Encountered error during initialization", e);
@@ -89,7 +99,9 @@ public class MetadbApp implements CommandLineRunner {
                     System.err.printf("\nCaught CTRL-C, shutting down gracefully...\n");
                     try {
                         requestReplyHandlingService.shutdown();
-                        messageHandlingService.shutdown();
+                        researchMessageHandlingService.shutdown();
+                        clinicalMessageHandlingService.shutdown();
+                        patientCorrectionHandlingService.shutdown();
                         messagingGateway.shutdown();
                     } catch (Exception e) {
                         LOG.error("Encountered error during shutdown process", e);
